Added function HasEffectApplyAuraName to check if a spell/aura has effect SPELL_EFFECT_APPLY_AURA and effect apply aura name x. * Function Aura::DotCanCrit() breaks if no aura with effect SPELL_AURA_ALLOW_DOT_TO_CRIT can be found for the caster. The functions should be checked out as soon as possible. * Function Unit::FindAuraWithAuraEffect loops threw all auras of the unit. Returns nullptr if the requested aura is not found in m_auras[340].
Warn compare equal types in MySQLDataStore.

+++ src/world/Storage/MySQLDataStore.cpp
@@ -2505,7 +2505,7 @@ void MySQLDataStore::LoadPlayerXpToLevelTable()
     _playerXPperLevelStore.clear();
     _playerXPperLevelStore.resize(sWorld.m_levelCap);
 
-    for (uint8 level = 0; level < sWorld.m_levelCap; ++level)
+    for (uint32 level = 0; level < sWorld.m_levelCap; ++level)
         _playerXPperLevelStore[level] = 0;
 
     QueryResult* player_xp_to_level_result = WorldDatabase.Query("SELECT player_lvl, next_lvl_req_xp FROM player_xp_for_level");
